Rename the lib
